+++ src/mesh/edge_measures.rs
@@ -19,7 +19,7 @@ impl Mesh
     ///
     /// Returns the length of the specified edge
     ///
-    pub fn edge_length(&self, halfedge_id: HalfEdgeID) -> f32
+    pub fn edge_length(&self, halfedge_id: HalfEdgeID) -> f64
     {
         let (p0, p1) = self.edge_positions(halfedge_id);
         (p0 - p1).magnitude()
@@ -28,7 +28,7 @@ impl Mesh
     ///
     /// Returns the squared length of the specified edge
     ///
-    pub fn edge_sqr_length(&self, halfedge_id: HalfEdgeID) -> f32
+    pub fn edge_sqr_length(&self, halfedge_id: HalfEdgeID) -> f64
     {
         let (p0, p1) = self.edge_positions(halfedge_id);
         (p0 - p1).magnitude2()
